modifiez la route DELETE

+++ backend/controllers/stuff.js
@@ -24,7 +24,7 @@ exports.modifyThing = (req, res, next) => {
   delete thingObject._userId;
   Thing.findOne({ _id: req.params.id})
     .then(thing => {
-      if (thing.userId != req.auth.userId) {
+      if (thing.userId !== req.auth.userId) {
         res.status(401).json({ message: 'Non-authorisé'})
       } else {
         Thing.updateOne({ _id: req.params.id}, { ...thingObject, _id: req.params.id })
@@ -36,9 +36,20 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-    .catch(error => res.status(404).json({ error }));
+  Thing.findOne({ _id: req.params.id})
+    .then(thing => {
+      if (thing.userId !== req.auth.userId) {
+        res.status(401).json({ message: 'Non-authorisé'})
+      } else {
+        const filename = thing.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Thing.deleteOne({ _id: req.params.id})
+            .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
+            .catch(error => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch( error => res.status(500).json({ error }));
 };
 
 exports.getOneThing = (req, res, next) => {
